+++ src/main/java/com/prac/ktb/auth/config/JwtProvider.java
@@ -1,11 +1,10 @@
-package com.prac.ktb.auth;
+package com.prac.ktb.auth.config;
 
 import io.jsonwebtoken.*;
 import io.jsonwebtoken.security.Keys;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-import org.springframework.stereotype.Service;
 
 import java.security.Key;
 import java.util.Base64;
