+++ src/main/java/com/prac/ktb/auth/controller/AuthController.java
@@ -1,0 +1,10 @@
+package com.prac.ktb.auth.controller;
+
+import jakarta.transaction.Transactional;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController("/auth/sessions")
+@Transactional
+public class AuthController {
+
+}
