+++ src/main/java/com/prac/ktb/auth/service/AuthService.java
@@ -1,0 +1,4 @@
+package com.prac.ktb.auth.service;
+
+public class AuthService {
+}
