--- src/main/java/com/prac/ktb/controller/TestController.java
@@ -1,14 +1,0 @@
-package com.prac.ktb.controller;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class TestController {
-
-    @GetMapping("/")
-    public ResponseEntity<String> test() {
-        return ResponseEntity.ok("성공");
-    }
-}
