+++ src/main/java/com/prac/ktb/user/service/UserService.java
@@ -1,7 +1,7 @@
-package com.prac.ktb.service;
+package com.prac.ktb.user.service;
 
-import com.prac.ktb.dto.UserRequestDto;
-import com.prac.ktb.entity.User;
+import com.prac.ktb.user.dto.UserRequestDto;
+import com.prac.ktb.user.entity.User;
 
 public interface UserService {
     User createUser(UserRequestDto userRequestDto);
